添加PC端业务代码

+++ health-management-client/src/main/java/com/yins/health/interceptor/LoginInterceptor.java
@@ -59,16 +59,16 @@ public class LoginInterceptor implements HandlerInterceptor {
                 log.info("使用特殊API密钥访问，跳过验证");
                 // 设置预定义的用户信息到ThreadLocal
                 AccountDto accountDTO = AccountDto.builder()
-                    .id(Long.valueOf(remoteServiceConfig.getSpecialId()) )
-                    .username("boss")
-                    .build();
+                        .id(Integer.valueOf(remoteServiceConfig.getSpecialId()) )
+                        .username("boss")
+                        .build();
                 threadLocal.set(accountDTO);
                 return true;
             }
 
             // 5. 解析 Token 并设置用户信息
             JWT jwt = JWTUtil.parseToken(token);
-            Long userId = Long.parseLong(jwt.getPayload("accountId").toString());
+            Integer userId = Integer.parseInt(jwt.getPayload("accountId").toString());
             // 3. 验证 Token 有效性
             //String rawRedisToken = stringRedisTemplate.opsForValue().get(RedisKeyEnum.LOGIN_TOKEN.getPrefix()+tokenKey);
             String tokenKey = redisKey.token_key_pre + userId;
